refactor: config 파일 분리
