+++ src/main/java/techcourse/myblog/web/config/ArgumentResolverConfig.java
@@ -1,0 +1,23 @@
+package techcourse.myblog.web.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import techcourse.myblog.web.support.LoginUserArgumentResolver;
+
+import java.util.List;
+
+@Configuration
+public class ArgumentResolverConfig implements WebMvcConfigurer {
+    @Bean
+    public LoginUserArgumentResolver loginUserMethodArgumentResolver() {
+        LoginUserArgumentResolver bean = new LoginUserArgumentResolver();
+        return bean;
+    }
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> argumentResolvers) {
+        argumentResolvers.add(loginUserMethodArgumentResolver());
+    }
+}
