+++ src/main/java/techcourse/myblog/web/config/CommonInterceptorConfig.java
@@ -1,0 +1,23 @@
+package techcourse.myblog.web.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.HandlerInterceptor;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import techcourse.myblog.web.interceptor.CommonInterceptor;
+
+@Configuration
+public class CommonInterceptorConfig implements WebMvcConfigurer {
+    private final HandlerInterceptor commonInterceptor;
+
+    public CommonInterceptorConfig(final CommonInterceptor commonInterceptor) {
+        this.commonInterceptor = commonInterceptor;
+    }
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(commonInterceptor)
+                .addPathPatterns("/**")
+                .excludePathPatterns("/logout");
+    }
+}
